# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clivvy API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "summary": "List Projects",
        "operationId": "list_projects_api_v1_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectView"
                  },
                  "title": "Response List Projects Api V1 Projects Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload",
        "operationId": "upload_api_v1_projects_post",
        "parameters": [
          {
            "name": "filename",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "summary": "Project Detail",
        "operationId": "project_detail_api_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove Queued",
        "operationId": "remove_queued_api_v1_projects__project_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/link": {
      "post": {
        "summary": "Import Link",
        "operationId": "import_link_api_v1_projects_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/selection": {
      "patch": {
        "summary": "Selection",
        "operationId": "selection_api_v1_projects__project_id__selection_patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipSelection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/retry": {
      "post": {
        "summary": "Retry",
        "operationId": "retry_api_v1_projects__project_id__retry_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/media/{kind}": {
      "get": {
        "summary": "Media",
        "operationId": "media_api_v1_projects__project_id__media__kind__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClipSelection": {
        "properties": {
          "start_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "End Ms"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "video_edits": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoEdits"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "start_ms",
          "end_ms",
          "revision"
        ],
        "title": "ClipSelection"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LinkImport": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 10,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "LinkImport"
      },
      "ProjectView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "has_audio": {
            "type": "boolean",
            "title": "Has Audio"
          },
          "start_ms": {
            "type": "integer",
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "detected_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Language"
          },
          "transcript_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript Language"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          },
          "stage": {
            "type": "string",
            "title": "Stage",
            "default": "Waiting for worker"
          },
          "shorts_count": {
            "type": "integer",
            "title": "Shorts Count",
            "default": 0
          },
          "video_edits": {
            "$ref": "#/components/schemas/VideoEdits"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "size_bytes",
          "status",
          "error",
          "duration_seconds",
          "width",
          "height",
          "has_audio",
          "start_ms",
          "end_ms",
          "revision"
        ],
        "title": "ProjectView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoEdits": {
        "properties": {
          "framing": {
            "type": "string",
            "enum": [
              "vertical",
              "horizontal"
            ],
            "title": "Framing",
            "default": "vertical"
          },
          "fit": {
            "type": "string",
            "enum": [
              "fit",
              "fill"
            ],
            "title": "Fit",
            "default": "fit"
          },
          "trim_start_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Trim Start Ms",
            "default": 0
          },
          "trim_end_ms": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trim End Ms"
          },
          "brightness": {
            "type": "number",
            "maximum": 0.3,
            "minimum": -0.3,
            "title": "Brightness",
            "default": 0
          },
          "contrast": {
            "type": "number",
            "maximum": 1.5,
            "minimum": 0.5,
            "title": "Contrast",
            "default": 1
          },
          "saturation": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Saturation",
            "default": 1
          },
          "speed": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.5,
            "title": "Speed",
            "default": 1
          },
          "volume": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Volume",
            "default": 1
          },
          "fade_in": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Fade In",
            "default": 0
          },
          "fade_out": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Fade Out",
            "default": 0
          },
          "rotation": {
            "type": "integer",
            "enum": [
              0,
              90,
              180,
              270
            ],
            "title": "Rotation",
            "default": 0
          },
          "flip": {
            "type": "boolean",
            "title": "Flip",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "VideoEdits",
        "description": "Small, reversible metadata; never accept arbitrary FFmpeg expressions."
      }
    }
  }
}